Update patch set 1
Patch Set 1:

> > No, I don't believe you did. All you mentioned was GPU decoding, and in this case for SLI or the like, this is tunable via BParameters.
> > 
> > IMFMediaSource in Windows Media Foundation has a single flag to enable or disable GPU decoding (actually the flag is for the entire pipeline.) If it works for them, I don't know why it would not work for us.
> 
> You are mixing two things there. The resulting media source in media2 will be indeed able to do that. But that's the IPC stuff. We are discussing of the low level API. How do you think that it is done internally? I guess also windows has something very similar to the codec kit. The fact that windows does it this way I think is not an argument.

Would you like me to dig out the MLT api documentation? Or Apple's CoreMedia API? My argument here is that "this is cleaner, and Windows/etc. do it and it works, so let's do that and not something more complex." Not that your system will not "work", but that in many other ways it will be worse than what I believe I am suggesting, which I think is better on all fronts (at least in this case it's also mostly the BMediaFile status quo.)

Further, the "IPC stuff" is not so complicated in those systems. Media Kit as it is overcomplicates this also. That's a different question though.

> If you add support for filters to BMediaFile it will come out something very similar to the BMediaExtractor, so you are not giving me a good point...

I have said BMediaFile should not have filters. That should be part of a graph node API indeed. If MediaKit cannot do that well now, then we need to design Media Kit 2, not introduce an alternative filter system in a different kit that will compete with Media Kit.

> > > > 24 stereo tracks @ 192kHz 32-bit float is 35MB/s. That's 21GB for 10 minutes, uncompressed, indeed. But video is exponentially more than that, and Blender's media graph system can handle it just fine, as can Adobe's, and these work primarily in RAM and only read the data off the disk, afaik.
> > > 
> > > Didn't you reply yourself? Do I have more to add?
> > 
> > I just said these systems don't do disk caching. Further for realtime audio, caching is totally irrelevant, as there is nothing to cache, only new data to process. So again your remarks don't make much sense to me...
> 
> Can we please concentrate on the issue?

Uh, do you remember why we started on this thread in the first place? It is because you suggested "what about caching; then the BDataIO that BMediaFile uses will not be the one you passed in, but a BPositionIO that wraps it for caching purposes." I said I could not ever think of a scenario where such a cache should be used, you then replied with your "multi GB" example, and then I replied as you can see above explaining why caching does not make any sense in that scenario.

So that was entirely "on topic" for this discussion, I think.

Patch-set: 1