Update patch set 1
Patch Set 1:

> > Haiku is designed around pervasive multi-threading. Locking as long as there is little contention is cheap; and context switching is also cheap. So this is a problem because of the complexity of setting it up, but not so because of the locking or context switches, necessarily. Doing away with those would be nice but not required.
> 
> As said, and I'm sorry for saying that, I think you miss very basic concepts of computer science.

That is a personal attack, not a reply. Which concept have I missed here?

I specifically said in the next paragraph that this system works but is indeed hard to program, and should be redesigned. So, again, what have I missed.

> > Of course your chain is more efficient. I am not claiming it isn't; I am claiming that you are creating a new graph node system that will compete with the Media Kit one. I already said that we should probably invent a Media Kit 2 with a better graph node system; but inventing a Codec Kit to make reading files "simpler" when in actuality you are creating a lot of complexity under the hood is kind of ridiculous.
> 
> Again, the codec kit *is not* a new graph node system. It just will provide a way to stack up modules (like it's done now anyway, considering it still a private API).

Stacking up modules is exactly what a graph node system is... What exactly do you think a graph node system is, then, if you claim the Codec Kit is not one?

> It's not going to be complex but about the same lines of code as BMediaFile with the difference that if you want, you can do more interesting things.

Yes, and doing more interesting things, like filters, is the responsibility of a media node framework...

> No. That's why I say that you talk without knowing the thing.

OK, then as requested above, please link me to some explanation instead of just telling me I am wrong.

> > I am not saying to merge them. BAdapterIO is and should remain the I/O handle. I don't think there should be a BStreamer though; the existing plugin API already did OK with sniffing for the most part. So BStreamer would be entirely superfluous as far as I can see.
> 
> What? There are two different sniffing processes involved. Again I'm not willing to discuss with you if you don't at least read the code. That's why I think you are trolling me...

I read the code. I guess the two sniffing procedures you are referring to are as follows:
 * sniffing the provided URL, to see what can access it
 * sniffing the media at the given URL, to see if we can decode it

The first should be done by the BAdapterIO implementation plugins (e.g. HTTPStreamer). The second should be done by the demuxer/decoder (e.g. FFmpeg plugin.) So where does BStreamer fit in there? It seems to be some sort of "mixed metaphor", partially the BAdapterIO and partially the demuxer/decoder?

Patch-set: 1