Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> It's not a computer science term, that's what I'm saying. I'm not saying it doesn't describe what BeOS do. But from a computer science perspective there's no such model as "pervasive multithreading" like for example we can define cooperative multitasking vs preemptive multitasking.

Fair enough, you are right. In the context of the discussion I think waddlesplash's point was that the 9 vs 4 thread may not be as much of a detriment on Haiku as other systems, because of the "pervasive multi-threading".

In the end I think we all agree the current node system is too complicated, but probably big fixes will need to wait for R2. I am quite sure many of your ideas will be used then.

Patch-set: 1