Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > It's not a computer science term, that's what I'm saying. I'm not saying it doesn't describe what BeOS do. But from a computer science perspective there's no such model as "pervasive multithreading" like for example we can define cooperative multitasking vs preemptive multitasking.
> 
> Fair enough, you are right. In the context of the discussion I think waddlesplash's point was that the 9 vs 4 thread may not be as much of a detriment on Haiku as other systems, because of the "pervasive multi-threading".

I did some tests, and I can ensure you this is not true nowadays anymore. There's a difference in how the linux scheduler works, that should be probably tuned.

> 
> In the end I think we all agree the current node system is too complicated, but probably big fixes will need to wait for R2. I am quite sure many of your ideas will be used then.

That this API should stay private until R2 may be an argument. There's also a misunderstanding here. It was decided that subsequent betas should have been forked from the previous one, and so the R1. So it was decided in multiple occasions, that once the beta branch has been forked people was able to beging working on R2 functionality, because someone (I think PulkoMandy said he wanted this role) would have merged changes back. Now it seems something changed, but I can't tell where and when it was decided, certainly I didn't express my opinion. 

But I'd like to hear from you, about what we are discussing now, if there's some usefulness or not in this API. What do you think?

Patch-set: 1