Update patch set 1
Patch Set 1:

> > That is a personal attack, not a reply. Which concept have I missed here?
> 
> The fact that in reality spawning an insane number of threads is not at a free cost.

app_server has a thread per window, and one BWindow is a thread, so that's 2 threads per window. In point of fact, one thread takes less than a hundred us to spawn, and if one is going to spend multiple minutes doing media processing, the time to spawn it is negligible. Runtime overhead, thanks to our O(1) scheduler, is also negligible.

> > I specifically said in the next paragraph that this system works but is indeed hard to program, and should be redesigned. So, again, what have I missed.
> 
> You continue to argue that using an insane number of threads is OK.

Yes, it's not optimal, but it is OK. It is hard to program and should be redesigned, but it is still, for now, OK.

> > > It's not going to be complex but about the same lines of code as BMediaFile with the difference that if you want, you can do more interesting things.
> > 
> > Yes, and doing more interesting things, like filters, is the responsibility of a media node framework...
> 
> This is your opinion. It seems to me that Axel agreed with me for example that you do not need to spawn a media node to use a video filter.

Only because spawning media nodes is a rather annoying task because of how tedious the media kit is to program, yes. So if we designed a new graph API, then we would not have this problem...

> > > No. That's why I say that you talk without knowing the thing.
> > 
> > OK, then as requested above, please link me to some explanation instead of just telling me I am wrong.
> 
> No. You are attacking with all those 1-min-and-bang comments. I'm worried of that.

So if I waited 10 minutes before posting them, that would somehow be different?

If the technical content of my comments is flawed, then explain how. Otherwise, just saying "you are writing your comments too quickly" is not a real criticism, it's just an ad hominem used to misdirect.

> > I read the code. I guess the two sniffing procedures you are referring to are as follows:
> >  * sniffing the provided URL, to see what can access it
> >  * sniffing the media at the given URL, to see if we can decode it
> > 
> > The first should be done by the BAdapterIO implementation plugins (e.g. HTTPStreamer). The second should be done by the demuxer/decoder (e.g. FFmpeg plugin.) So where does BStreamer fit in there? It seems to be some sort of "mixed metaphor", partially the BAdapterIO and partially the demuxer/decoder?
> 
> So you are saying that BStreamer and BAdapterIO should be merged, just like I asked before. You retracted. Now come out again with this. Clearly you didn't read the code or you troll me...

I listed 2 bullet points above for the 2 types of sniffing I see happening in your code. Is there some 3rd type I missed? Otherwise, what other distinction am I missing?

I am not "trolling" you, I am trying to have an actual technical discussion here, and you are not replying with any real answers, only statements like "it won't work", "you are wrong", "that is stupid", etc. These are not answers to questions.

Patch-set: 1