Update patch set 1
Patch Set 1:

> > Is my summary incorrect?
> 
> Let's suppose it is a superfluous functionality, then what? That being said it's not superfluous, it allows to use custom classes for buffering. The caching example is just an example.

Then my summary was not incorrect, and your "What!? Are you serious?" makes even less sense than before.

> That this API should stay private until R2 may be an argument. There's also a misunderstanding here. It was decided that subsequent betas should have been forked from the previous one, and so the R1.

No, as far as I am aware, this was never decided. It was strongly advocated for by PulkoMandy, and I equally as strongly argued against it, partially on the grounds that it would fragment development efforts and progress would be slower than ever, and R1 would likely never get out then. In the end it seems we have gone with my suggestion, partly because nobody seemed to want to spend the effort working on PulkoMandy's.

> So it was decided in multiple occasions, that once the beta branch has been forked people was able to beging working on R2 functionality, because someone (I think PulkoMandy said he wanted this role) would have merged changes back. Now it seems something changed, but I can't tell where and when it was decided, certainly I didn't express my opinion. 

It was never decided in the first place. As you can clearly see, bugfixes were not being merged back into r1beta1 besides certain crucial updates, while my XHCI or FreeBSD compat refactors remained in master only; and of course master still builds with GCC2, which is clear evidence we did not go with this route.

Patch-set: 1