Update patch set 2
Patch Set 2:

Hi,
What is the use for this? If the structure is not used anywhere in the code, why add it? And if you have some usage for it in mind, I think it should be included in the same commit.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0