Burl: rework handling of URL encoding
The content of a BUrl should always be in encoded form, to simplify
handling and validation. Remove the UrlEncode member function, and
add a new way to handle URL encoding:
- All ways to set an URL (constructors, SetUrlString, and all setters)
now take an extra boolean parameter indicating if the string is already
encoded. The default value is to encode strings automatically.
- The static version of UrlEncode and UrlDecode, which operate on a
  string, are preserved and used by other parts of the API.

All unit tests adjusted to handle this, and still passing.

Fixes #12983
Change-Id: I06f06978d0d35e56d7c92b67f001856bb7dcafc8