Update patch set 1
Patch Set 1:

> No, as far as I am aware, this was never decided. It was strongly advocated for by PulkoMandy, and I equally as strongly argued against it, partially on the grounds that it would fragment development efforts and progress would be slower than ever, and R1 would likely never get out then. In the end it seems we have gone with my suggestion, partly because nobody seemed to want to spend the effort working on PulkoMandy's.

Also because there was nothing merged in master that prevents using it as the base for beta2 (and R1) so far. This is why I did not put much effort on the beta1 branch. IIRC after discussing this at length, we had more or less agreed that there is no way to know what will happen, and that we would go with my solution if the other one does not prove to be workable.

So far, the master branch has remained rather stable since beta1, there was a lot of work focused on bugfixes. There are two a bit annoying changes that I can think of: my own addition of BLooper::Loop (ABI extension) that I want to use in new WebKit versions so we backported it to beta1, and Barrett refactoring work in media kit, which for examples makes it difficult to keep StreamRadio working on both master and beta branches (and I did not see an intention or request to backport this for beta). Both changes are quite minor (one is an ABI extension, and the other is mostly a problem because of moving things from one .h to another IIRC) so it seems fine to base beta2 on the current master branch, unless more invasive changes are made during the next 6 months.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0