Update patch set 2
Patch Set 2:

> I get that, but why a for loop? What are we iterating on here? On open firmware platform there will be only one architecture, so we should not need a loop.

We are iterating over boot *platforms*, not architectures. On x86_64 for instance there are 3: bios_ia32, pxe_ia32, and efi. You can theoretically use U-Boot to boot SPARC, right? So in the future there may be more than just OpenFirmware on SPARC, so we need the loop here too, even if for now it will only iterate once.

Patch-set: 2