[WIP] GSoC 2019: Implementing PCID in x86_64
This patch is a draft proposal for GSoC 2019 project.

NOTE: the code changes in this patch is rather primitive and might not even
      pass the compilation. It is just an illustration of the potential, and
      could change significantly in the future.

For now, the idea is to mimick linux for a dual paga table solution,
then apply PCID for acceleration. When PCID is enabled, we have the last
12 bits of CR3 register to specify a PCID number, so that writing to CR3
register no longer causes a full TLB flush.

For the dual page table part, we allocate an extra top-level PML4 page
for each X86VMTranslationMap64Bit, which contains just limited kernel
mappings. This translation map is per-process from my point of view, so
we are able to implement dual page table for every process.

To mitigate potential side-channel attack, we should allocate different
PCID for user-space and kernel-space tasks. Splitting the PCID into 2 parts
is feasible, one from [0 - 2047], and the other from [2048 - 4096]. Upon
each switch from user space to kernel space, we set the bit 11 to 0, and
when leaving kernel space to user space, we set the bit 11 to 1. This means
the kernel will always use PCID range [0 - 2047] and user space will get
the remaining [2048 - 4096].

Since PCID is limited, we might not want to allocate a new PCID for every
new process, as it would soon run out (maybe not in Haiku?). Instead, we
could still take a per-cpu strategy that each CPU could allow at most 6 PCID
at the same time (Although if the programs in Haiku would not reach to
thousand level, we could try to allocate new PCID for each program as well).

The code change regarding to this patch involves:
* Every syscall entry and interrupt entry specified by MSR
* Every possible place that will return to user space
* Every time we are creating a new translation map
* Every context switch when we need to switch to another thread

Please correct me if I misunderstant anything.

Signed-off-by: Zihan Yang <whois.zihan.yang@gmail.com>