Update patch set 2
Patch Set 2:

> Sorry for the inconvenience. When I tried picking this commit and squashing the previous one, I kept getting an error message :(

We are not going to merge it this way.

Squashing the two commits cannot result in an error message, I'm fairly sure git knows how to do that. You had problems with change-ids, but that may be related to the use of change-id 0 in the other commit. You can simply remove that (and, I think, reuse the change-id from this commit for the new squashed commit).

Patch-set: 2
Label: Code-Review=0