Update patch set 1
Patch Set 1:

> Further, why does the kernel need 2048 PCIDs? It only has one address space, can't it just use PCID 0?

The kernel accesses other process's memory, this can be implemented by having, for each process, an userspace+kernel mapping, and an userspace-only mapping. This way the kernel (when handling a syscall) can still access the current process memory easily.

The alternative would be a single space for the kernel, but then we would need to dynamically map in and out the parts of the relevant userspace process we want to access inside that single address space.

> And yes, getting to 4096 address spaces will be rare. We can probably allocate them per process and then swap out the "last used" one when we get over 4096.

4096 process is a little low limit, so it would be nice to have a plan for when we get past this. Allocating all IDs and sharing a few of them is an option, using one ID per CPU and adjusting them everytime we shedule a process is another. I don't know which one is better.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0