intel_extreme.accelerant: create display modes list from VBT using create_display_modes()
For cases when EDID is not available but VBT information was retrieved,
previous code generated display mode list manually, potentially
incorrectly, causing the black screen during boot. This commit
fixes this by using a dedicated function, used for the
case where EDID is available (or VBT is not available).

Fixes #14280.

Change-Id: I95cfc5313260f0b9a01a98ba78e300b4383b6e32