Update patch set 1
Patch Set 1:

A simple table would be enough:

struct search engine {
   const char* shortcut;
   const char* url;
};

search_engine engines[] = {
    { "g", "https://google.com/search?q=%s" }
}

A map allows a slightly faster search, but does not preserve the sorting, which is annoying if the user wants the list in a specific order in the preferences.

A BMessage allows to easily store it on disk (when we allow customization), but may not be the most efficient in-memory structure.

Patch-set: 1