Update patch set 6
Patch Set 6:

> @Adrien and Waddlesplash. What do you think? Should I do make changes as the Sergei sugessted?

According to our interface guidelines (https://www.haiku-os.org/docs/HIG/index.xml), an alert should be used only when there is no other way. Since we now have disabling of the window, the alert is not needed. A message in the window as in CodyCam (in addition to disabling the controls) will be enough.

Patch-set: 6