OS.h: Don't include pthread.h.
This was done for 1 rather rarely-used function, as pthread.h cannot
define the thread_id type, nor include OS.h, as this pollutes the namespace
unacceptably. But including this way also pollutes the namespace unacceptably,
as OS.h is used in kernel mode, also, where pthread.h cannot go.

Instead, just define the symbol with a return type of "int32". Since it
is a C symbol, this is perfectly acceptable.