Update patch set 1
Patch Set 1:

Yes, BLocker on BeOS allows for lock ownership transfer.

For example, this pattern is possible:
- Create a lock and lock it
- Create a thread
- In the thread, initialize some stuff, then unlock the lock when we are ready to process events

In other words, in BeOS, locks are not associated with a specific thread, and there are valid usage patterns for that.

Apparently we wanted to disallow this, adding a debugger() call. This allowed to find various applications and libraries using that pattern, which you can find in #6004 and duplicates.

We can decide to disallow this from now on, but it is a "soft" ABI change, especially as this usage pattern is documented in the Be Book. So, what is suggested here is to preserve the ABI for gcc2, and change the behavior for other architectures, making BLocker bound to a specific thread there (making the behavior similar to a pthread_mutex). In that case, we need a different way to implement the pattern mentionned above.

Patch-set: 1