Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I am struggling to think of a legitimate pattern of code where that kind of lock ownership flow would be a good idea; it implies a kind of memory ownership transfer I'm not sure would be good.

We do support this in our kernel locks for example. The example I gave is a good use case: you want to block things until a thread is started.

So
1) Create the lock and lock it
2) Spawn the thread
3) When the thread is ready to process events, it will unlock the lock

In fact I suspect BeOS did exactly this when starting a BLooper. The constructor creates it with the lock held, and it is unlocked when the thread starts running its event loop.

It is however incompatible with recursive locks (since the locking is not bound to a specific thread, you have no way to know wether a lock is recursive).

In POSIX, one would implement this using a semaphore instead, where ownership is not tracked (only waiting threads are). My guess is, in an attempt to simplify things, the BeAPI went with a solution that allows implementing both recursive locks and semaphore-like behavior, in a single class.

That was probably a bad idea, but we should still provide some semaphore-like solution in the BeAPI as well, so there is a suggested replacement for this for the few apps that use it. We probably have a benaphore implementation that would do the job just fine?

Patch-set: 1