Update patch set 1
Patch Set 1:

As already mentionned, this use case is explicitly allowed in the be book:

"Any thread can call Unlock() on any BLocker—the thread needn't be the lock's current holder. Call IsLocked() before calling Unlock() if you want to make sure you own a lock before you unlock it."

https://www.haiku-os.org/legacy-docs/bebook/BLocker.html#BLocker_Lock

So it is usable both as a recursive lock and as a transferrable one, which means no error checking can be made.

In the kernel, the transfer from one thread to another has to be done explicitly, see http://xref.plausible.coop/source/xref/haiku/headers/private/kernel/lock.h#273 (recursive_lock_transfer_lock). That's more sane, as it allows to do the error checks, but it's still a strange thing to do with a recursive lock, hence the function checks that the lock has only 1 level of locking before allowing it.

And indeed this is used exactly once in our codebase, and it doesn't work, see #15015.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0