Update patch set 1
Patch Set 1:

> Patch Set 1:
...
> Why is there a restriction on the language list at all? Shouldn't users be able to pick whichever language they want?
> I don't know where the account language is used, but at least for user ratings I see no reason for this restriction.

Hello Adrien; Are you able to clarify where you see the restriction?  I am not sure I understand what you mean.  The user is able to specify a 'preferred language' when they create an HDS user (for use in the web interface) and are also able to specify the language for any given user-rating at the time that they create the user-rating.  The list of languages is now determined by HDS rather than a hard-coded list of a few languages in HaikuDepot.

Patch-set: 1