Update patch set 1
Patch Set 1:

> Hello Adrien; Are you able to clarify where you see the restriction?  I am not sure I understand what you mean.  The user is able to specify a 'preferred language' when they create an HDS user (for use in the web interface) and are also able to specify the language for any given user-rating at the time that they create the user-rating.  The list of languages is now determined by HDS rather than a hard-coded list of a few languages in HaikuDepot.

I am wondering why this is the responsibility of HDS, as we have a list of languages available in the locale kit already. Why do we need to rely on the server for this? What additional restrictions does it apply to the list of languages in the locale kit?

Patch-set: 1