Update patch set 1
Patch Set 1:

> Patch Set 1:
> I am wondering why this is the responsibility of HDS, as we have a
> list of languages available in the locale kit already. Why do we
> need to rely on the server for this? What additional restrictions
> does it apply to the list of languages in the locale kit?

The data for the user-rating is ultimately stored in the relational database in the HDS server.  In the database schema [1] (strangely not rendering too well) you will see that the NaturalLanguage is related to the UserRating.  Owing to this relationship, there is a referential integrity check.  Because the list of languages in the Haiku operating system might differ from the set of the languages in the HDS haikudepot.natural_language table, it is important that the list of the languages offered to the user in the Haiku Depot desktop application is sourced from the server.  This is important because if the natural language code were not able to be found on the server then the record would not be able to be stored and the API request would fail returning an error to the desktop application.

[1] https://depot.haiku-os.org/__docs/datamodel.html

Patch-set: 1
Label: Code-Review=0