Update patch set 1
Patch Set 1:

There is a good reason we are using a fork of an old version of glibc. It is known to expose some non-POSIX symbols, which BeOS applications link against.

This is why I have already mentionned that I'm fine with replacing the libc, but not for gcc2. You have ignored this.

If you want to show that it is not a problem, you will have to use abi-compliance-checker or a similar tool to show us that indeed, no symbols are changed, and therefore the ABI is still the same, or if there are changes, we will need to investigate them to make sure they are not putting a risk to the ABI.

Also, we are supposed to keep a stable ABI since we are in beta stage, so it seems to not be the right time to do this.

Patch-set: 1
Label: Code-Review=0