Update patch set 1
Patch Set 1:

> There is a good reason we are using a fork of an old version of glibc. It is known to expose some non-POSIX symbols, which BeOS applications link against.
> 
> This is why I have already mentionned that I'm fine with replacing the libc, but not for gcc2. You have ignored this.

I have not removed any APIs with this very concern in mind.

> If you want to show that it is not a problem, you will have to use abi-compliance-checker or a similar tool to show us that indeed, no symbols are changed, and therefore the ABI is still the same, or if there are changes, we will need to investigate them to make sure they are not putting a risk to the ABI.

Yes, korli already requested this, I'll be sure to do it.

> Also, we are supposed to keep a stable ABI since we are in beta stage, so it seems to not be the right time to do this.

We have already said that we are not bug compatibile with BeOS, and our malloc is already one major break with that. This will not change ABI at all if done properly, which is why I'm taking great care in doing it.

Patch-set: 1