Update patch set 1
Patch Set 1:

Some notes:

I didn't add support for unloading or live change, because it's tricky to know when we can safely unload the addon. Since some thread might still be in a call inside the old addon after we change the pointer.

One possible way would be to lock/unlock all known loopers in sequence after the pointer is set, to make sure all drawing code is exited (since most if not all) calls to the object are drawing code and require a locked looper. Then invalidate all windows to force a redraw. Not sure how it'd impact the layout.

I only pass around the path to the add-on, but maybe we just want the name stored in the settings? (but then where do we load it from? We'd need to search all locations for it.)

Patch-set: 1