Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Looks good, and makes sense, but I couldn't find a link where this is documented.

The POSIX documentation at opengroup.org (IEEE Std 1003.1-2017 (Revision of IEEE Std 1003.1-2008)) doesn't really appear to indicate anything as far as the descriptor being required to be open with write perms, nor do any of the defined errors, so I'm assuming this is simply a case of aligning it with behavior of other platforms? On a side note, is it intentional that the copyright of the attached test is assigned to you rather than Jerome?

Patch-set: 1
Reviewer: Gerrit User 1000015 <1000015@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0