Update patch set 2
Patch Set 2:

> Well I reviewed some of the musl code and I'm going to -1 this for now because it's ugly and unreadable. Single letter variable names, obscure for loops, complete lack of comments.

It looks very much like the FreeBSD code does, actually. Haiku's code is unique in how clean and readable it is.

The point is not so much to pick the most readable implementation for the C library, but to pick one which is the best written (musl far outclasses glibc here especially in security-related matters), at least relatively well tested (musl is used as the default libc on a number of Linux distros, including Alpine Linux).

> Also, except one file, this is only replacing code from BSD, NewOS or our own, which is already under an acceptable license. So nothing is gained in that regard.

Yes, because these components are the easiest to begin by replacing. Currently we have our own implementations because glibc's were inadequate or nonexistent in the glibc version we use. The eventual goal is to have as much of the libc as possible be from one source, not a variety of sources as we currently have, so these need to be replaced also.

Patch-set: 2