Update patch set 3
Patch Set 3:

(4 comments)

> Patch Set 2:
> I would really like to avoid hardcoded printf. I don't mind keeping them active for now
...

I have removed that specific log case to get this resolved for now, but there are other cases in the application.  Some years ago I added a 'log level' command-line flag that the user can use at start time to control the level of logging.  The usage is "-v off|info|debug|trace".  Especially with the server interactions, I have found this to be enormously helpful to casually get additional debugging from people and much easier then trying to get them to obtain or create a debug build of the same version.  Are you happy with such a "printf" as long as it is constrained by the checks for log level?  An example of this can be seen at "LocalPkgDataLoadProcess.cpp:115".

Patch-set: 3
Label: Code-Review=0