Update patch set 2
Patch Set 2:

I don't agree here. Our code readability is what makes it possible for people to onboard Haiku and start contributing. If our libc code is clean and readable, people can investigate it and fix issues in it.

So, I agree we could do something to replace the messy floating point support we inherited from glibc: it's quite messy and something you don't really want to dig into. There, investigating musl as a replacement is fine (assuming there are no regressions and no ABI changes, which is far from trivial and may result in even more hacks).

But, I don't agree we should replace clean and perfectly working parts of the codebase, especially as these are quite simple functions which have 0 maintenance cost. They were written 10 or 15 years ago and untouched since.

Also, thinking we are going to use musl and lower our maintenance costs is not exactly true. We're going to have to keep up to date with upstream, make sure it still builds with gcc2, and maintain at least the buildsystem (jamfiles vs whatever they use). Unless we make libc a separate package built at haikuports, our libc will be a fork of musl then. It will not be "just musl and we can report bugs to them when we have problems and let them do the work".

So, I see a clear loss of ease of development (less readable code, have to remember upstreaming changes to them and getting their bugfixes), and so far I fail to see what we win (I can't say our C library is unmaintained or creating problems, especially in this area).

Patch-set: 2