Update patch set 2
Patch Set 2:

> I don't agree here. Our code readability is what makes it possible for people to onboard Haiku and start contributing. If our libc code is clean and readable, people can investigate it and fix issues in it.

The point of using musl is so that we have a robust, well-tested libc that people don't need to modify. Indeed, many of these files from our implementation are cleaner; but when is the last time anyone contributed to, or even modified those?

We are trying to reduce the maintenance burden of our libc here by using another implementation. Trading some cleanliness for robustness and completeness is alright with me in the libc, anyway. There are too many moving parts here that I'd rather someone else worry about.

If in the future we have an overabundance of contributors, then maybe we can worry about writing our own custom libc. But I really don't see the need for that right now.

> So, I agree we could do something to replace the messy floating point support we inherited from glibc: it's quite messy and something you don't really want to dig into. There, investigating musl as a replacement is fine (assuming there are no regressions and no ABI changes, which is far from trivial and may result in even more hacks).

As I have repeatedly stated, the C library APIs have not changed aside from GNU extensions, and the C ABI has not changed either, which is what allows us to use GCC7 FFmpeg on GCC2.

> But, I don't agree we should replace clean and perfectly working parts of the codebase, especially as these are quite simple functions which have 0 maintenance cost. They were written 10 or 15 years ago and untouched since.

The point here is to have as much of our libc as possible from one source for consistency.

> Also, thinking we are going to use musl and lower our maintenance costs is not exactly true. We're going to have to keep up to date with upstream, make sure it still builds with gcc2, and maintain at least the buildsystem (jamfiles vs whatever they use). Unless we make libc a separate package built at haikuports, our libc will be a fork of musl then. It will not be "just musl and we can report bugs to them when we have problems and let them do the work".

They use makefiles which are pretty clean. The jamfiles are not going to be much more than file lists. So that is not a maintenance burden.

Keeping up to date with upstream is also a pretty easy task, as there aren't going to be that many changes to make. musl is mostly C89, I've only had to modify these files very lightly in spots.

It may be a fork, but it's definitely a soft one.

> So, I see a clear loss of ease of development (less readable code, have to remember upstreaming changes to them and getting their bugfixes), and so far I fail to see what we win (I can't say our C library is unmaintained or creating problems, especially in this area).

In the area covered by these files specifically, no. But in libio and math, it absolutely is (fish crashing, pow() problems, etc.)

Patch-set: 2